requirements.txt
appnope==0.1.4
asn1crypto==1.5.1
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
ipykernel==6.29.3
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.5.0
lxml==5.2.2
matplotlib-inline==0.1.7
multitasking==0.0.11
nest_asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
peewee==3.17.5
pexpect==4.9.0
pg8000==1.31.2
pickleshare==0.7.5
pip==24.0
platformdirs==4.2.2
prompt-toolkit==3.0.42
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.32.2
scramp==1.4.5
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.2
tornado==6.3.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
yfinance==0.2.40
zipp==3.17.0
dbt-core==1.7.10
dbt-snowflake==1.7
